Add a test to get the ball rolling
