+++ Source/Drosky/Drosky.swift
@@ -23,53 +23,6 @@ import Deferred
  */
 
 
-// Mark: HTTP method and parameter encoding
-
-public enum HTTPMethod: String {
-    case GET, POST, PUT, DELETE, OPTIONS, HEAD, PATCH, TRACE, CONNECT
-}
-
-extension HTTPMethod {
-    func alamofireMethod() -> Alamofire.HTTPMethod {
-        switch self {
-        case .GET:
-            return .get
-        case .POST:
-            return .post
-        case .PUT:
-            return .put
-        case .DELETE:
-            return .delete
-        case .OPTIONS:
-            return .options
-        case .HEAD:
-            return .head
-        case .PATCH:
-            return .patch
-        case .TRACE:
-            return .trace
-        case .CONNECT:
-            return .connect
-        }
-    }
-}
-
-public enum HTTPParameterEncoding {
-    case url
-    case json
-}
-
-extension HTTPParameterEncoding {
-    func alamofireParameterEncoding() -> Alamofire.ParameterEncoding {
-        switch self {
-        case .url:
-            return URLEncoding.default
-        case .json:
-            return JSONEncoding.default
-        }
-    }
-}
-
 // MARK:- DroskyResponse
 
 public struct DroskyResponse {
@@ -98,36 +51,6 @@ extension DroskyResponse: CustomStringConvertible {
     }
 }
 
-
-// MARK:- Router
-
-public typealias Signature = (header: String, value: String)
-
-struct Router {
-    let environment: Environment
-    let signature: Signature?
-    
-    func urlRequest(forEndpoint endpoint: Endpoint) -> TaskResult<URLRequestConvertible> {
-        guard let URL = URL(string: environment.routeURL(endpoint.path)) else {
-            return .failure(DroskyErrorKind.malformedURLError)
-        }
-        
-        var request = URLRequest(url: URL)
-        request.httpMethod = endpoint.method.alamofireMethod().rawValue
-        request.allHTTPHeaderFields = endpoint.httpHeaderFields
-        if let signature = self.signature {
-            request.setValue(signature.value, forHTTPHeaderField: signature.header)
-        }
-        
-        do {
-            let alamofireEncoding = endpoint.parameterEncoding.alamofireParameterEncoding()
-            let request = try alamofireEncoding.encode(request, with: endpoint.parameters)
-            return .success(request)
-        } catch let error {
-            return .failure(error)
-        }
-    }
-}
 
 // MARK: - Drosky
